+++ lib/routes/i-cable/templates/description.art
@@ -1,0 +1,8 @@
+{{ if media.length > 0 }}
+    {{ each media }}
+        <figure><img src="{{ $value.source_url }}" /></figure>
+    {{ /each }}  
+{{ /if }}
+{{ if content }}
+    {{@ content }}
+{{ /if }}
